nem_weather_price_ml/requirements.txt
pandas>=2.0.0
pyarrow>=14.0.0
PyYAML>=6.0
scikit-learn>=1.3.0
streamlit>=1.28.0
flask>=3.0.0
